Added abilities to let astroSynth Simulate sparsly sampled data
Added alias frequency function to POS to allow for alias's to be introduced into data. Also allowed for progress bars to be disabled module wide at instantiation time. Made the behavior of this disabling consient across all methds. Additionally added the ability to query the LSP of an entire ephemieris in POS